LD fixing small formatting forgets, changing category for tag and committing the home page
