+++ app/loading.tsx
@@ -1,4 +1,3 @@
-
 import LoadingComponent from '@/components/loading/loading'
 
 export default function Loading () {
