+++ components/RecentSalesByCategory/RecentSalesByCategory.tsx
@@ -7,15 +7,15 @@ interface Sale {
   id: number;
   price: number;
   date: Date;
-  fromUser: string;
-  toUser: string;
+  fromPlayer: string;
+  toPlayer: string;
   nft: {
     issuedId: number;
     item: {
       name: string;
       imageUrl: string;
       rarityName: string;
-      categories: string[];
+      tags: string[];
     };
   };
 }
@@ -26,17 +26,16 @@ interface Category {
 
 interface Props {
   data: Sale[];
-  categories: Category[];
+  tags: Category[];
 }
 
-const RecentSalesByCategory: React.FC<Props> = ({ data, categories })  => {
+const RecentSalesByCategory: React.FC<Props> = ({ data, tags })  => {
   const [selectedOption, setSelectedOption] = useState("all");
 
   const filteredData =
     selectedOption === "all"
       ? data
-      : data.filter((item) => item.nft.item.categories.includes(selectedOption));
-
+      : data.filter((item) => item.nft.item.tags.includes(selectedOption));
 
   const firstFiveItems = filteredData.slice(0, 5)
 
@@ -44,49 +43,46 @@ const RecentSalesByCategory: React.FC<Props> = ({ data, categories })  => {
     <div className={styles["latest-sales-container"]}>
       <div className={styles["sales-header"]}>
         <div className={styles["sales-title"]}>RecentSales</div>
-
         <div className={styles["radios-container"]}>
-        <div
-              className={styles["input-and-label-container"]}
-            >
-              <input
-                type="radio"
-                value="all"
-                id="all"
-                checked={selectedOption === "all"}
-                onChange={() => setSelectedOption("all")}
-              />
-              <label htmlFor="all"> All </label>
-            </div>
-          {categories.map((category, index) => (
+          <div className={styles["input-and-label-container"]}>
+            <input
+              type="radio"
+              value="all"
+              id="all"
+              checked={selectedOption === "all"}
+              onChange={() => setSelectedOption("all")}
+            />
+            <label htmlFor="all"> All </label>
+          </div>
+          {/* {tags.map((tag, index) => (
             <div
               key={index}
               className={styles["input-and-label-container"]}
             >
               <input
                 type="radio"
-                value={category.name}
-                id={category.name}
-                checked={selectedOption === category.name}
-                onChange={() => setSelectedOption(category.name)}
+                value={tag.name}
+                id={tag.name}
+                checked={selectedOption === tag.name}
+                onChange={() => setSelectedOption(tag.name)}
               />
-              <label htmlFor={category.name}>{category.name}</label>
+              <label htmlFor={tag.name}>{tag.name}</label>
             </div>
-          ))}
+          ))} */}
         </div>
       </div>
       <div className={styles["cards-container"]}>
         {firstFiveItems.map((sale, index) => (
           <RecentSaleCard
-          key={index}
-          price={sale.price}
-          date={sale.date}
-          toUser={sale.toUser}
-          fromUser={sale.fromUser}
-          issuedId={sale.nft.issuedId}
-          itemName={sale.nft.item.name}
-          imageUrl={sale.nft.item.imageUrl}
-        />
+            key={index}
+            price={sale.price}
+            date={sale.date}
+            toPlayer={sale.toPlayer}
+            fromPlayer={sale.fromPlayer}
+            issuedId={sale.nft.issuedId}
+            itemName={sale.nft.item.name}
+            imageUrl={sale.nft.item.imageUrl}
+          />
         ))}
       </div>
     </div>
