--- lib/prisma/categories.ts
@@ -1,14 +1,0 @@
-import prisma from "./prisma";
-
-export async function getCategories() {
-    try {
-        const categories = await prisma.category.findMany()
-
-return categories
-
-    } catch (error) {
-      console.error("Error:", error);
-    } finally {
-      await prisma.$disconnect();
-    }
-  }
